[Koa] Move to Koa / simpler error handling and response creation / improved queries
Moved to Koa which is a trimmed-down version of Express that has generator support built in. As part of this is provides a DOM-like capture-bubble traversal of the middleware stack that is pretty easy to use. This lets us consolidate error handling and response formatting to ensure that all responses have an "err" field for example.

Refactored the code into separate route files, etc. so the API endpoints are separate from the web page endpoints.

Simplified the ReQL query for checking/marking used codes so that there's just none query.