Add messaging for sunsetting Digitimate
We're planning on sunsetting Digitimate gracefully and want developers using it to smoothly migrate to another service. First we'll communicate to developers about this change and give them at least 180 days to plan. This commit adds visible messaging to the GitHub README and the homepage to start raising awareness.

We'll also need to build an API endpoint for developers to export data for the developer email address they used.